Moving files btw packages
* NopHit to openaria.core.format.nop
* NopEncoder to openaria.core.format.nop
* NopToCsv to openaria.core.package
