+++ open-aria-airborne/src/main/java/org/mitre/openaria/RunAirborneViaStdIn.java
@@ -12,7 +12,7 @@ import java.util.Iterator;
 import org.mitre.openaria.airborne.AirbornePairConsumer;
 import org.mitre.openaria.core.Point;
 import org.mitre.openaria.core.PointIterator;
-import org.mitre.openaria.core.formats.NopHit;
+import org.mitre.openaria.core.formats.nop.NopHit;
 import org.mitre.openaria.core.formats.nop.NopParser;
 import org.mitre.openaria.system.StreamingKpi;
