+++ open-aria-airborne/src/test/java/org/mitre/openaria/airborne/SingleAircraftRecordTest.java
@@ -14,7 +14,7 @@ import org.mitre.caasd.commons.DataCleaner;
 import org.mitre.caasd.commons.Speed;
 import org.mitre.openaria.core.Track;
 import org.mitre.openaria.core.TrackPair;
-import org.mitre.openaria.core.formats.NopHit;
+import org.mitre.openaria.core.formats.nop.NopHit;
 
 import org.junit.jupiter.api.Test;
