+++ open-aria-core/src/main/java/org/mitre/openaria/smoothing/CoastedPointRemover.java
@@ -9,8 +9,8 @@ import java.util.TreeSet;
 import org.mitre.caasd.commons.DataCleaner;
 import org.mitre.openaria.core.Point;
 import org.mitre.openaria.core.Track;
-import org.mitre.openaria.core.formats.NopHit;
 import org.mitre.openaria.core.formats.nop.CenterRadarHit;
+import org.mitre.openaria.core.formats.nop.NopHit;
 import org.mitre.openaria.core.formats.nop.StarsRadarHit;
 
 public class CoastedPointRemover<T> implements DataCleaner<Track<T>> {
