+++ open-aria-core/src/test/java/org/mitre/openaria/core/InterpolateTest.java
@@ -11,7 +11,7 @@ import java.time.Instant;
 import org.mitre.caasd.commons.Distance;
 import org.mitre.caasd.commons.LatLong;
 import org.mitre.caasd.commons.Spherical;
-import org.mitre.openaria.core.formats.NopHit;
+import org.mitre.openaria.core.formats.nop.NopHit;
 
 import org.junit.jupiter.api.Test;
