+++ open-aria-core/src/test/java/org/mitre/openaria/core/PointsTest.java
@@ -27,7 +27,7 @@ import java.util.stream.Collectors;
 import org.mitre.caasd.commons.Distance;
 import org.mitre.caasd.commons.LatLong;
 import org.mitre.caasd.commons.TimeWindow;
-import org.mitre.openaria.core.formats.NopHit;
+import org.mitre.openaria.core.formats.nop.NopHit;
 
 import org.junit.jupiter.api.Test;
