+++ open-aria-core/src/test/java/org/mitre/openaria/smoothing/RemoveLowVariabilityTracksTest.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 
 import org.mitre.openaria.core.Point;
 import org.mitre.openaria.core.Track;
-import org.mitre.openaria.core.formats.NopHit;
+import org.mitre.openaria.core.formats.nop.NopHit;
 
 import org.junit.jupiter.api.Test;
