+++ open-aria-core/src/test/java/org/mitre/openaria/smoothing/RemoveTracksMatchingPredicateTest.java
@@ -8,7 +8,7 @@ import java.util.function.Consumer;
 
 import org.mitre.openaria.core.Point;
 import org.mitre.openaria.core.Track;
-import org.mitre.openaria.core.formats.NopHit;
+import org.mitre.openaria.core.formats.nop.NopHit;
 
 import org.junit.jupiter.api.Test;
