+++ open-aria-core/src/test/java/org/mitre/openaria/smoothing/StarsSmoothingTest.java
@@ -4,7 +4,7 @@ package org.mitre.openaria.smoothing;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import org.mitre.openaria.core.formats.NopHit;
+import org.mitre.openaria.core.formats.nop.NopHit;
 import org.mitre.openaria.core.formats.nop.NopMessageType;
 import org.mitre.openaria.core.formats.nop.StarsRadarHit;
