+++ open-aria-core/src/test/java/org/mitre/openaria/smoothing/TrimLowSpeedGroundPointsTest.java
@@ -14,7 +14,7 @@ import org.mitre.caasd.commons.Distance;
 import org.mitre.caasd.commons.Speed;
 import org.mitre.openaria.core.Point;
 import org.mitre.openaria.core.Track;
-import org.mitre.openaria.core.formats.NopHit;
+import org.mitre.openaria.core.formats.nop.NopHit;
 import org.mitre.openaria.core.formats.nop.NopParsingUtils;
 
 import org.junit.jupiter.api.Test;
