+++ open-aria-core/src/test/java/org/mitre/openaria/smoothing/VerticalOutlierDetectorTest.java
@@ -15,8 +15,8 @@ import java.util.stream.Stream;
 
 import org.mitre.openaria.core.Point;
 import org.mitre.openaria.core.Track;
-import org.mitre.openaria.core.formats.NopEncoder;
-import org.mitre.openaria.core.formats.NopHit;
+import org.mitre.openaria.core.formats.nop.NopEncoder;
+import org.mitre.openaria.core.formats.nop.NopHit;
 import org.mitre.openaria.smoothing.VerticalOutlierDetector.AnalysisResult;
 
 import org.junit.jupiter.api.Test;
