+++ open-aria-system/src/main/java/org/mitre/openaria/system/KafkaIngestor.java
@@ -5,8 +5,8 @@ import static com.google.common.collect.Lists.newArrayList;
 import static java.util.Objects.requireNonNull;
 import static org.apache.kafka.clients.consumer.ConsumerConfig.PARTITION_ASSIGNMENT_STRATEGY_CONFIG;
 import static org.mitre.caasd.commons.util.PropertyUtils.*;
-import static org.mitre.openaria.core.formats.NopHit.parseSafely;
 import static org.mitre.openaria.core.formats.nop.Facility.toFacility;
+import static org.mitre.openaria.core.formats.nop.NopHit.parseSafely;
 import static org.mitre.openaria.core.formats.nop.NopMessageType.isNopRadarHit;
 import static org.mitre.openaria.kafka.KafkaPropertyUtils.verifyKafkaBrokers;
 
@@ -25,8 +25,8 @@ import java.util.stream.IntStream;
 import org.mitre.caasd.commons.util.ErrorCatchingTask;
 import org.mitre.caasd.commons.util.ExceptionHandler;
 import org.mitre.openaria.core.Point;
-import org.mitre.openaria.core.formats.NopHit;
 import org.mitre.openaria.core.formats.nop.Facility;
+import org.mitre.openaria.core.formats.nop.NopHit;
 import org.mitre.openaria.core.temp.Extras.HasSourceDetails;
 import org.mitre.openaria.kafka.PartitionMapping;
 import org.mitre.openaria.system.tools.DataLatencySummarizer;
