+++ open-aria-threading/src/test/java/org/mitre/openaria/threading/TrackMakerTest.java
@@ -15,7 +15,7 @@ import org.mitre.caasd.commons.ConsumingCollections.ConsumingArrayList;
 import org.mitre.openaria.core.Point;
 import org.mitre.openaria.core.PointBuilder;
 import org.mitre.openaria.core.Track;
-import org.mitre.openaria.core.formats.NopHit;
+import org.mitre.openaria.core.formats.nop.NopHit;
 
 import org.junit.jupiter.api.Test;
